Render with prev until it's gone, default to that
The thing is we have the new indices before we've pushed that to the
gpu, so if we add meshes and try to draw with the old the counts are
off. I am so glad I figured this out. I just thought about it and kept
triggering it until my brain made the connection